refactor: move login function to container
